requirements.txt
black==23.3.0
Brotli==1.0.9
crispy-bootstrap4==2022.1
Django==4.1.7
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.1
django-tables2==2.5.3
djhtml==3.0.6
gunicorn==20.1.0
libsass==0.22.0
unpoly==0.3.0
Werkzeug==2.2.3
whitenoise==6.4.0